fix: pass files as buffer
