+++ lib/manager/gradle-wrapper/artifacts.spec.ts
@@ -3,8 +3,9 @@ import { resolve } from 'path';
 import Git from 'simple-git/promise';
 import * as dcUpdate from '.';
 import { platform as _platform } from '../../platform';
-import { mocked } from '../../../test/util';
+import { mocked, getName } from '../../../test/util';
 import { ifSystemSupportsGradle } from '../gradle/__testutil__/gradle';
+import { setFsConfig } from '../../util/fs';
 
 const platform = mocked(_platform);
 const config = {
@@ -23,14 +24,14 @@ async function resetTestFiles() {
   });
 }
 
-describe('manager/gradle-wrapper/update', () => {
+describe(getName(__filename), () => {
+  beforeAll(() => {
+    setFsConfig(config);
+  });
+  beforeEach(() => {
+    jest.resetAllMocks();
+  });
   describe('updateArtifacts - replaces existing value', () => {
-    beforeEach(() => {
-      jest.resetModules();
-      jest.resetAllMocks();
-      jest.clearAllMocks();
-    });
-
     ifSystemSupportsGradle(6).it('replaces existing value', async () => {
       try {
         jest.setTimeout(5 * 60 * 1000);
@@ -64,7 +65,6 @@ describe('manager/gradle-wrapper/update', () => {
             'gradlew.bat',
           ].map(fileProjectPath => {
             return {
-              artifactError: null,
               file: {
                 name: fileProjectPath,
                 contents: readFileSync(
@@ -98,12 +98,6 @@ describe('manager/gradle-wrapper/update', () => {
     });
   });
   describe('updateArtifacts - up to date', () => {
-    beforeEach(() => {
-      jest.resetModules();
-      jest.resetAllMocks();
-      jest.clearAllMocks();
-    });
-
     ifSystemSupportsGradle(6).it('up to date', async () => {
       try {
         jest.setTimeout(5 * 60 * 1000);
@@ -148,12 +142,6 @@ describe('manager/gradle-wrapper/update', () => {
     });
   });
   describe('updateArtifacts - error handling - getRepoStatus', () => {
-    beforeEach(() => {
-      jest.resetModules();
-      jest.resetAllMocks();
-      jest.clearAllMocks();
-    });
-
     ifSystemSupportsGradle(6).it('error handling - getRepoStatus', async () => {
       try {
         jest.setTimeout(5 * 60 * 1000);
@@ -201,12 +189,6 @@ describe('manager/gradle-wrapper/update', () => {
   });
 
   describe('updateArtifacts - error handling - command execution', () => {
-    beforeEach(() => {
-      jest.resetModules();
-      jest.resetAllMocks();
-      jest.clearAllMocks();
-    });
-
     ifSystemSupportsGradle(6).it(
       'error handling - command execution',
       async () => {
