+++ lib/workers/branch/index.ts
@@ -1,8 +1,6 @@
 import { DateTime } from 'luxon';
-import { readFile } from 'fs-extra';
 import is from '@sindresorhus/is';
 import minimatch from 'minimatch';
-import { join } from 'upath';
 import { logger } from '../../logger';
 import { isScheduledNow } from './schedule';
 import { getUpdatedPackageFiles } from './get-updated';
@@ -40,6 +38,7 @@ import {
 import { BranchStatus } from '../../types';
 import { exec } from '../../util/exec';
 import { regEx } from '../../util/regex';
+import { readLocalFile } from '../../util/fs';
 
 // TODO: proper typings
 function rebaseCheck(config: RenovateConfig, branchPr: any): boolean {
@@ -367,12 +366,10 @@ export async function processBranch(
                 { file: relativePath, pattern },
                 'Post-upgrade file saved'
               );
-              const existingContent = await readFile(
-                join(config.localDir, relativePath)
-              );
+              const existingContent = await readLocalFile(relativePath);
               config.updatedArtifacts.push({
                 name: relativePath,
-                contents: existingContent.toString(),
+                contents: existingContent,
               });
             }
           }
